+++ sshd-common/src/main/java/org/apache/sshd/common/util/threads/ExecutorServiceProvider.java
@@ -19,13 +19,22 @@
 
 package org.apache.sshd.common.util.threads;
 
+import java.util.function.Supplier;
+
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 @FunctionalInterface
-public interface ExecutorServiceCarrier {
+public interface ExecutorServiceProvider {
     /**
-     * @return The {@link CloseableExecutorService} to use
+     * @return A {@link Supplier} of {@link CloseableExecutorService} to be used when
+     * asynchronous execution required. If {@code null} then a single-threaded
+     * ad-hoc service is used.
      */
-    CloseableExecutorService getExecutorService();
+    Supplier<? extends CloseableExecutorService> getExecutorServiceProvider();
+
+    default CloseableExecutorService resolveExecutorService() {
+        Supplier<? extends CloseableExecutorService> provider = getExecutorServiceProvider();
+        return (provider == null) ? null : provider.get();
+    }
 }