+++ sshd-sftp/src/main/java/org/apache/sshd/server/subsystem/sftp/SftpSubsystemFactory.java
@@ -27,6 +27,7 @@ import org.apache.sshd.common.subsystem.sftp.SftpConstants;
 import org.apache.sshd.common.util.GenericUtils;
 import org.apache.sshd.common.util.ObjectBuilder;
 import org.apache.sshd.common.util.threads.CloseableExecutorService;
+import org.apache.sshd.common.util.threads.ManagedExecutorServiceSupplier;
 import org.apache.sshd.server.channel.ChannelSession;
 import org.apache.sshd.server.command.Command;
 import org.apache.sshd.server.subsystem.SubsystemFactory;
@@ -36,7 +37,8 @@ import org.apache.sshd.server.subsystem.SubsystemFactory;
  */
 public class SftpSubsystemFactory
         extends AbstractSftpEventListenerManager
-        implements SubsystemFactory, SftpEventListenerManager, SftpFileSystemAccessorManager {
+        implements ManagedExecutorServiceSupplier, SubsystemFactory,
+        SftpEventListenerManager, SftpFileSystemAccessorManager {
 
     public static final String NAME = SftpConstants.SFTP_SUBSYSTEM_NAME;
     public static final UnsupportedAttributePolicy DEFAULT_POLICY = UnsupportedAttributePolicy.Warn;
@@ -97,16 +99,14 @@ public class SftpSubsystemFactory
         return NAME;
     }
 
+    @Override
     public Supplier<? extends CloseableExecutorService> getExecutorServiceProvider() {
         return executorsProvider;
     }
 
-    /**
-     * @param provider The {@link Supplier} of {@link CloseableExecutorService}-s to be used by the
-     * {@link SftpSubsystem} command when starting execution. If {@code null} then a single-threaded
-     * ad-hoc service is used.
-     */
-    public void setExecutorServiceProvider(Supplier<? extends CloseableExecutorService> provider) {
+    @Override
+    public void setExecutorServiceProvider(
+            Supplier<? extends CloseableExecutorService> provider) {
         executorsProvider = provider;
     }
 
@@ -138,11 +138,6 @@ public class SftpSubsystemFactory
 
     public void setErrorStatusDataHandler(SftpErrorStatusDataHandler handler) {
         errorStatusDataHandler = Objects.requireNonNull(handler, "No error status data handler provided");
-    }
-
-    protected CloseableExecutorService resolveExecutorService() {
-        Supplier<? extends CloseableExecutorService> provider = getExecutorServiceProvider();
-        return (provider == null) ? null : provider.get();
     }
 
     @Override